+++ Sources/KeyboardKit/_Deprecated/KeyboardLayout+Deprecated.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public extension KeyboardLayout {
+    
+    @available(*, deprecated, renamed: "hasKeyboardSwitcher(_:)")
+    func hasKeyboardSwitcher(for type: Keyboard.KeyboardType) -> Bool {
+        itemRows.hasKeyboardSwitcher(type)
+    }
+    
+}
+
+@available(*, deprecated, renamed: "KeyboardLayoutRowIdentifiable")
+public typealias KeyboardLayoutRowItem = KeyboardLayoutRowIdentifiable