+++ Sources/KeyboardKit/_Deprecated/KeyboardLayoutRowItem.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "KeyboardLayoutRowIdentifiable")
-public typealias KeyboardLayoutRowItem = KeyboardLayoutRowIdentifiable